# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for Sonit Server",
        "title": "Sonit Server API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/product-inventory-transactions/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new product inventory transaction record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product-inventory-transactions"
                ],
                "summary": "Create product inventory transaction",
                "parameters": [
                    {
                        "description": "Create product inventory transaction request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateProductInventoryTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid data. Please try again.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "401": {
                        "description": "You have no rights to access this action.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "500": {
                        "description": "There is something wrong in the system during the process. Please try again later.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    }
                }
            }
        },
        "/product-inventory-transactions/product/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves inventory transactions filtered by product criteria",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product-inventory-transactions"
                ],
                "summary": "Get inventory transactions by product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page_number",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search keyword",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter property (e.g. date, price)",
                        "name": "filter_prop",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order (ASC or DESC)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Inventory action (e.g. import, export, sale, return)",
                        "name": "action",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.PaginationDataResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid data. Please try again.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "401": {
                        "description": "You have no rights to access this action.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "500": {
                        "description": "There is something wrong in the system during the process. Please try again later.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    }
                }
            }
        },
        "/product-inventory-transactions/update": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates an existing product inventory transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product-inventory-transactions"
                ],
                "summary": "Update product inventory transaction",
                "parameters": [
                    {
                        "description": "Update product inventory transaction request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateProductInventoryTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid data. Please try again.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "401": {
                        "description": "You have no rights to access this action.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "500": {
                        "description": "There is something wrong in the system during the process. Please try again later.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    }
                }
            }
        },
        "/product-inventory-transactions/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a specific product inventory transaction by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product-inventory-transactions"
                ],
                "summary": "Get product inventory transaction by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product inventory transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.ProductInventoryTransaction"
                        }
                    },
                    "400": {
                        "description": "Invalid data. Please try again.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "401": {
                        "description": "You have no rights to access this action.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "500": {
                        "description": "There is something wrong in the system during the process. Please try again later.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.ProductInventoryTransaction": {
            "type": "object",
            "properties": {
                "action": {
                    "description": "Import, Export, Sale",
                    "type": "string"
                },
                "amount": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "request.CreateProductInventoryTransactionRequest": {
            "type": "object",
            "required": [
                "action",
                "amount",
                "date",
                "product_id"
            ],
            "properties": {
                "action": {
                    "description": "Import, Export, Sale",
                    "type": "string"
                },
                "amount": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                }
            }
        },
...
        "request.UpdateProductInventoryTransactionRequest": {
            "type": "object",
            "required": [
                "transaction_id"
            ],
            "properties": {
                "transaction_id": {
                    "type": "string"
                },
                "update_data": {
                    "$ref": "#/definitions/request.CreateProductInventoryTransactionRequest"
                }
            }
        },
...
        "response.MessageAPIResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "response.PaginationDataResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "page_number": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}